Function rust_algorithms::dfs::search_node

pub fn search_node(
    tree: Tree, 
    start_node_id: isize, 
    search_node_id: isize
) -> Option<SearchResult>

A function to search for the path to a node.

This function takes two node ids, for the start and target node, and computes a path between them. This path consists of a Vec of Links.

Example:

use rust_algorithms::graph::*;
use rust_algorithms::dfs::*;
let mut tree = Tree::new();
let mut node1 = Node::new("Node 1");
let mut node2 = Node::new("Node 2");
node1.id = tree.add_node(node1);
node2.id = tree.add_node(node2);
let link1 = Link::new((node1.id, node2.id), 5);
tree.add_link(link1);
let result = search_node(tree, node1.id, node2.id).unwrap();
let link0 = Link::new((node1.id, node1.id), 0);
assert_eq!(result.links[0], link0);
assert_eq!(result.links[1], link1);
assert_eq!(result.cost, 5);